requirements.txt
altair==5.5.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.55.1
fuzzywuzzy==0.18.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
Levenshtein==0.26.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
narwhals==1.15.2
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.0.0
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-Levenshtein==0.26.1
pytz==2024.2
RapidFuzz==3.11.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.40.2
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.45.1
